% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sp_bars.R
\name{add_sp_bars}
\alias{add_sp_bars}
\title{Add Mean ± Error Bars to a SuperPlot}
\usage{
add_sp_bars(
  p,
  bars,
  df,
  cond,
  rep_summary,
  bar_params = list(),
  crossbar_params = list()
)
}
\arguments{
\item{p}{ggplot object}

\item{bars}{character string, one of "mean_sd", "mean_sem", or "mean_ci"}

\item{df}{data frame of summary data (supplied by superplot())}

\item{cond}{character string of the condition column}

\item{rep_summary}{character string of the replicate summary column}

\item{bar_params}{named list of additional parameters passed to the
errorbar \code{stat_summary} layer}

\item{crossbar_params}{named list of additional parameters passed to the
crossbar \code{stat_summary} layer}
}
\value{
ggplot object
}
\description{
Add Mean ± Error Bars to a SuperPlot
}
