% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superplot_spec.R
\name{autoplot.superplot_spec}
\alias{autoplot.superplot_spec}
\title{Plot a SuperPlot Specification}
\usage{
\method{autoplot}{superplot_spec}(object, ...)
}
\arguments{
\item{object}{object returned by superplot_spec}

\item{...}{additional arguments (currently unused)}
}
\value{
ggplot object
}
\description{
Plot a SuperPlot Specification
}
