% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_alpha}
\alias{check_alpha}
\title{Check alpha argument}
\usage{
check_alpha(arg)
}
\arguments{
\item{arg}{argument passed as alpha}
}
\value{
none
}
\description{
Check alpha argument
}
\keyword{internal}
