% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_bars}
\alias{check_bars}
\title{Check bars argument}
\usage{
check_bars(arg)
}
\arguments{
\item{arg}{argument passed as bars}
}
\value{
none
}
\description{
Check bars argument
}
\keyword{internal}
