% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_colour}
\alias{check_colour}
\title{Check colour argument}
\usage{
check_colour(arg)
}
\arguments{
\item{arg}{argument passed as colour}
}
\value{
none
}
\description{
Check colour argument
}
\keyword{internal}
