% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_datadist}
\alias{check_datadist}
\title{Check datadist argument}
\usage{
check_datadist(arg)
}
\arguments{
\item{arg}{argument passed as datadist}
}
\value{
none
}
\description{
Check datadist argument
}
\keyword{internal}
