% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_fsize}
\alias{check_fsize}
\title{Check fsize argument}
\usage{
check_fsize(arg)
}
\arguments{
\item{arg}{argument passed as fsize}
}
\value{
none
}
\description{
Check fsize argument
}
\keyword{internal}
