% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_info}
\alias{check_info}
\title{Check info argument}
\usage{
check_info(arg)
}
\arguments{
\item{arg}{argument passed as info}
}
\value{
none
}
\description{
Check info argument
}
\keyword{internal}
