% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_linking}
\alias{check_linking}
\title{Check linking argument}
\usage{
check_linking(arg)
}
\arguments{
\item{arg}{argument passed as linking}
}
\value{
none
}
\description{
Check linking argument
}
\keyword{internal}
