% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_rep_summary}
\alias{check_rep_summary}
\title{Check rep_summary argument}
\usage{
check_rep_summary(arg)
}
\arguments{
\item{arg}{argument passed as rep_summary}
}
\value{
none
}
\description{
Check rep_summary argument
}
\keyword{internal}
