% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_shapes}
\alias{check_shapes}
\title{Check shapes argument}
\usage{
check_shapes(arg)
}
\arguments{
\item{arg}{argument passed as shapes}
}
\value{
none
}
\description{
Check shapes argument
}
\keyword{internal}
