% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_size}
\alias{check_size}
\title{Check size argument}
\usage{
check_size(arg)
}
\arguments{
\item{arg}{argument passed as size}
}
\value{
none
}
\description{
Check size argument
}
\keyword{internal}
