% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_stats}
\alias{check_stats}
\title{Check stats argument}
\usage{
check_stats(arg)
}
\arguments{
\item{arg}{argument passed as stats}
}
\value{
none
}
\description{
Check stats argument
}
\keyword{internal}
