% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_stats_test}
\alias{check_stats_test}
\title{Check stats_test argument}
\usage{
check_stats_test(arg)
}
\arguments{
\item{arg}{argument passed as stats_test}
}
\value{
none
}
\description{
Check stats_test argument
}
\keyword{internal}
