% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_xlab}
\alias{check_xlab}
\title{Check xlab argument}
\usage{
check_xlab(arg)
}
\arguments{
\item{arg}{argument passed as xlab}
}
\value{
none
}
\description{
Check xlab argument
}
\keyword{internal}
