% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_ylab}
\alias{check_ylab}
\title{Check ylab argument}
\usage{
check_ylab(arg)
}
\arguments{
\item{arg}{argument passed as ylab}
}
\value{
none
}
\description{
Check ylab argument
}
\keyword{internal}
