% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatplot.R
\name{flatplot}
\alias{flatplot}
\title{Make a FlatPlot}
\usage{
flatplot(
  df,
  meas,
  cond,
  colour = "#000000",
  xlab = "",
  ylab = "Measurement",
  datadist = "sina",
  size = 2,
  alpha = 0.5,
  bars = "mean_sd",
  fsize = 12,
  gg = NULL,
  stats = FALSE,
  stats_test = "para_unpaired"
)
}
\arguments{
\item{df}{data frame with at least three columns: meas, cond, repl}

\item{meas}{character name of column with measurement (e.g. intensity)}

\item{cond}{character name of column with condition (e.g. Control, WT)}

\item{colour}{string for colour palette to use, select ("rl_green", "rl_red",
"rl_blue", "rl_purple", "rl_orange", "rl_magenta", or a hex colour, default
is black)}

\item{xlab}{string for x label (default is empty)}

\item{ylab}{string for y label (default is "Measurement")}

\item{datadist}{string for data distribution to use, select ("sina" default,
or "jitter")}

\item{size}{numeric size of data points (default is 2)}

\item{alpha}{numeric vector of alpha range data and summary points (default
is c(0.5, 0.7))}

\item{bars}{string for type of error bars to add, select ("none",
"mean_sd" (default), "mean_sem", or "mean_ci")}

\item{fsize}{numeric font size for text (default is 12)}

\item{gg}{ggplot object to add to (default is NULL)}

\item{stats}{logical for whether to add statistical tests (default is FALSE)}

\item{stats_test}{string for statistical test to use, select
("para_unpaired", "para_paired", "nonpara_unpaired", or "nonpara_paired")}
}
\value{
ggplot object
}
\description{
Make a FlatPlot
}
\examples{
flatplot(lord_jcb, "Speed", "Treatment", ylab = "Speed (um/min)")

}
