% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fp_colour.R
\name{get_fp_colour}
\alias{get_fp_colour}
\alias{get_fp_color}
\title{FlatPlot Colour Selection (Single Colour)}
\usage{
get_fp_colour(key)

get_fp_color(key)
}
\arguments{
\item{key}{character name, can be "rl_green", "rl_red", "rl_blue",
"rl_yellow", "rl_purple", "rl_orange", "rl_magenta", "rl_inv", or a hex
colour, or a valid R color name}
}
\value{
character of hex colour
}
\description{
This function takes a single character name/code and returns the
corresponding hex colour. It checks if the input is a valid hex colour code
or a valid R color name, and if not, it uses a predefined mapping for
specific color names like "rl_green", "rl_red", etc. If the input is not
valid, it will return NULL. Use of \code{get_fp_color} is valid too.
}
\examples{
get_fp_colour("rl_green")
get_fp_color("rl_green")
}
