% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_info.R
\name{get_sp_info}
\alias{get_sp_info}
\title{Information about a SuperPlot}
\usage{
get_sp_info(
  df,
  meas,
  cond,
  repl,
  pal,
  xlab,
  ylab,
  datadist,
  size,
  alpha,
  bars,
  linking,
  rep_summary,
  shapes,
  fsize,
  gg,
  stats,
  stats_test,
  ...
)
}
\arguments{
\item{df}{data frame with at least three columns: meas, cond, repl}

\item{meas}{character name of column with measurement (e.g. intensity)}

\item{cond}{character name of column with condition (e.g. Control, WT)}

\item{repl}{character name of column with replicate (e.g. unique experiment
identifiers)}

\item{pal}{argument passed to pal}

\item{xlab}{argument passed to xlab}

\item{ylab}{argument passed to ylab}

\item{datadist}{argument passed to datadist}

\item{size}{argument passed to size}

\item{alpha}{argument passed to alpha}

\item{bars}{argument passed to bars}

\item{linking}{argument passed to linking}

\item{rep_summary}{argument passed to rep_summary}

\item{shapes}{argument passed to shapes}

\item{fsize}{argument passed to fsize}

\item{gg}{argument passed to gg}

\item{stats}{argument passed to stats}

\item{stats_test}{argument passed to stats_test}

\item{...}{additional arguments}
}
\value{
none
}
\description{
This function prints information about a SuperPlot, it runs internally from
superplot()
}
\keyword{internal}
