% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_shapes.R
\name{get_sp_shapes}
\alias{get_sp_shapes}
\title{Get shapes for data and summary plots}
\usage{
get_sp_shapes(n, shape_logical)
}
\arguments{
\item{n}{number of shapes to return}

\item{shape_logical}{logical for whether to use different shapes for
replicates. Default is FALSE, which results in a circle being used. TRUE
means that a selection of shapes are used, with data and summary points for
each replicate being the same. Four shapes are possible, so if n is greater
than 4, the shapes will repeat (but with different colors).}
}
\value{
contiguous vector of shape integers
}
\description{
Get shapes for data and summary plots
}
\examples{
get_sp_shapes(3, FALSE)
}
