% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_stats.R
\name{get_sp_stats}
\alias{get_sp_stats}
\title{Get statistical test results for the plotted data}
\usage{
get_sp_stats(df, rep_summary, cond, repl, ncond, nrepl, stats_test)
}
\arguments{
\item{df}{data frame containing the data}

\item{rep_summary}{summary statistic to use for testing}

\item{cond}{column name for the condition}

\item{repl}{column name for the replicate}

\item{ncond}{number of unique values in cond}

\item{nrepl}{number of unique values in repl}

\item{stats_test}{string for statistical test to use, select
("para_unpaired", "para_paired", "nonpara_unpaired", or "nonpara_paired")}
}
\value{
nothing, prints results to console
}
\description{
Get statistical test results for the plotted data
}
