% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{ignore_unused_imports}
\alias{ignore_unused_imports}
\title{Ignore unused imports}
\usage{
ignore_unused_imports()
}
\description{
Currently patchwork is used in the documentation but not in the package
itself. This function is a workaround to avoid warnings about unused imports
when building the package.
}
\keyword{keyword}
