% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lord_jcb}
\alias{lord_jcb}
\title{Example cell migration data}
\format{
\subsection{\code{lord_jcb}}{

A data frame with 300 rows and 3 columns:
\describe{
\item{Replicate}{Experimental replicate number (integer)}
\item{Treatment}{Conditions}
\item{Speed}{Measurement}
...
}
}
}
\source{
\url{https://doi.org/10.1083/jcb.202001064}
}
\usage{
lord_jcb
}
\description{
Data used in Lord et al 2020 to generate the plots in Figs S4 and S5
}
\keyword{datasets}
