% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pieplot.R
\name{pieplot}
\alias{pieplot}
\title{Make a pie chart using ggplot2}
\usage{
pieplot(x1, cols, x2 = NULL, label = NULL, ...)
}
\arguments{
\item{x1}{A numeric vector containing the values for the first
layer of the pie chart.}

\item{cols}{A character vector containing the colors for the
slices of the pie chart.}

\item{x2}{An optional numeric vector containing the values for the
second layer of the pie chart.}

\item{label}{An optional character string to be used as the title}

\item{...}{Currently not used.}
}
\value{
ggplot object
}
\description{
The function \code{pie_maker} creates a pie chart using \code{ggplot2}.
It takes two vectors of values and optionally a
second vector of values to create a pie chart with two layers.
}
\examples{
pieplot(x1 = c(123, 456),
  cols = c("#44aa99", "#117733"))

pieplot(x1 = c(50 - 20, 20, 80, 180 - 80),
  cols = c("#bbbbbb", "#44aa99", "#117733", "#dddddd"), x2 = c(100,130))

}
