% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representative.R
\name{representative}
\alias{representative}
\title{Report the representative datapoints in a SuperPlot}
\usage{
representative(
  df,
  meas,
  cond,
  repl,
  label = "",
  bars = "",
  rep_summary = "rep_mean",
  outlier = FALSE
)
}
\arguments{
\item{df}{data frame with at least three columns: meas, cond, repl}

\item{meas}{character name of column with measurement (e.g. intensity)}

\item{cond}{character name of column with condition (e.g. Control, WT)}

\item{repl}{character name of column with replicate (e.g. unique experiment
identifiers)}

\item{label}{character name of column with labels for measurements (e.g. file
names), optional.}

\item{bars}{string to specify the summary stats of replicate sumaries, select
("none" default, "mean_sd", "mean_sem", or "mean_ci"). Sepcifying this
overrides the \code{rep_summary} argument}

\item{rep_summary}{string for summary statistic to use for replicates, select
("rep_mean" default, or "rep_median")}

\item{outlier}{logical, if TRUE then ranking is reversed to find the most
extreme datapoint in each treatment and replicate, default is FALSE.}
}
\value{
data frame with the representative datapoints for each treatment and
replicate, with the rank of the difference from the summary statistic.
}
\description{
This function finds the representative datapoints for each condition and
replicate in a SuperPlot dataset. It calculates the mean or median of the
replicates (conditions), then ranks the individual measurements by their difference from
the summary statistic. The function returns a data frame with the
representative datapoints for each treatment and replicate, along with their
rank of difference from the summary statistic.
}
\details{
The top ranked datapoint for each treatment and replicate is printed to the
console. This is useful for identifying the most representative datapoint
for each condition, which can be used for showing in a figure.

A label column can be specified to identify the datapoint (this could be a
file name or other identifier). If no label is specified, the row number is
used instead.
}
\examples{
representative(lord_jcb,
  "Speed", "Treatment", "Replicate")

}
