% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_rproj.R
\name{setup_rproj}
\alias{setup_rproj}
\title{Ensure working directory setup}
\usage{
setup_rproj()
}
\value{
No return value
}
\description{
Checks if the directories ("Data", "Output", "Output/Data",
"Output/Plots", "Script") exist in the current working directory. If any of
these directories are missing, it creates them to ensure that the project
structure is properly set up.
}
