% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superplot_spec.R
\name{sp_modify}
\alias{sp_modify}
\title{Modify Advanced SuperPlot Options}
\usage{
sp_modify(spec, options = NULL, ...)
}
\arguments{
\item{spec}{object returned by superplot_spec}

\item{options}{named list of advanced options}

\item{...}{additional named options merged into the existing options}
}
\value{
updated superplot_spec object
}
\description{
Modify Advanced SuperPlot Options
}
