% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sp_bars.R
\name{trio_ci}
\alias{trio_ci}
\title{Calculate error bars (mean ± 95\% confidence interval)}
\usage{
trio_ci(x)
}
\arguments{
\item{x}{input from stat_summary}
}
\value{
data frame of y, ymin, and ymax
}
\description{
Calculate error bars (mean ± 95\% confidence interval)
}
\keyword{internal}
