% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sp_bars.R
\name{trio_mean}
\alias{trio_mean}
\title{Calculate mean cross bar}
\usage{
trio_mean(x)
}
\arguments{
\item{x}{input from stat_summary}
}
\value{
data frame of y, ymin, and ymax
}
\description{
Calculate mean cross bar
}
\keyword{internal}
