% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sp_bars.R
\name{trio_sem}
\alias{trio_sem}
\title{Calculate error bars (mean ± standard error of the mean)}
\usage{
trio_sem(x)
}
\arguments{
\item{x}{input from stat_summary}
}
\value{
data frame of y, ymin, and ymax
}
\description{
Calculate error bars (mean ± standard error of the mean)
}
\keyword{internal}
