% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_args}
\alias{validate_args}
\title{Validate SuperPlot Arguments}
\usage{
validate_args(
  pal = NULL,
  xlab = NULL,
  ylab = NULL,
  datadist = NULL,
  size = NULL,
  alpha = NULL,
  bars = NULL,
  linking = NULL,
  rep_summary = NULL,
  shapes = NULL,
  fsize = NULL,
  gg = NULL,
  stats = NULL,
  stats_test = NULL
)
}
\arguments{
\item{pal}{argument passed to pal}

\item{xlab}{argument passed to xlab}

\item{ylab}{argument passed to ylab}

\item{datadist}{argument passed to datadist}

\item{size}{argument passed to size}

\item{alpha}{argument passed to alpha}

\item{bars}{argument passed to bars}

\item{linking}{argument passed to linking}

\item{rep_summary}{argument passed to rep_summary}

\item{shapes}{argument passed to shapes}

\item{fsize}{argument passed to fsize}

\item{gg}{argument passed to gg}

\item{stats}{argument passed to stats}

\item{stats_test}{argument passed to stats_test}
}
\value{
none
}
\description{
Validate SuperPlot Arguments
}
\keyword{internal}
