% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify_fp_columns}
\alias{verify_fp_columns}
\title{Verify the data frame used for FlatPlot}
\usage{
verify_fp_columns(df, meas, cond)
}
\arguments{
\item{df}{data frame with at least three columns: meas, cond, repl}

\item{meas}{character name of column with measurement (e.g. intensity)}

\item{cond}{character name of column with condition (e.g. Control, WT)}
}
\value{
logical to allow plot to go ahead
}
\description{
Verify the data frame used for FlatPlot
}
\keyword{internal}
