% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fp_colour.R
\name{get_fp_colours}
\alias{get_fp_colours}
\title{FlatPlot Colour Selection (Multiple Colors)}
\usage{
get_fp_colours(keys)
}
\arguments{
\item{keys}{character vector of color names/codes, can be "rl_green", "rl_red", "rl_blue",
"rl_yellow", "rl_purple", "rl_orange", "rl_magenta", hex colours, or valid R color names}
}
\value{
character vector of hex colours
}
\description{
FlatPlot Colour Selection (Multiple Colors)
}
\examples{
get_fp_colours(c("rl_green", "#FF0000", "blue"))
get_fp_colours("rl_green")
}
