% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_colours.R
\name{get_sp_colours}
\alias{get_sp_colours}
\title{SuperPlot Colour Palette Selection}
\usage{
get_sp_colours(n, scheme)
}
\arguments{
\item{n}{integer number of colours to return}

\item{scheme}{character name, colour scheme to use
Choose from one of four palettes from Paul Tol ("tol_bright", "tol_vibrant",
"tol_muted", "tol_light"), or Color Universal Design ("cud"). Or a vector of
colours in hex format, e.g. c("#FF0000", "#00FF00", "#0000FF").}
}
\value{
character vector of colours
}
\description{
SuperPlot Colour Palette Selection
}
\examples{
get_sp_colours(3, "tol_bright")
}
