% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superplot.R
\name{superplot}
\alias{superplot}
\title{Make a SuperPlot}
\usage{
superplot(
  df,
  meas,
  cond,
  repl,
  pal = "tol_bright",
  xlab = "",
  ylab = "Measurement",
  datadist = "sina",
  size = c(2, 3),
  alpha = c(0.5, 0.7),
  bars = "",
  linking = FALSE,
  rep_summary = "rep_mean",
  shapes = FALSE,
  fsize = 12,
  gg = NULL
)
}
\arguments{
\item{df}{data frame with at least three columns: meas, cond, repl}

\item{meas}{character name of column with measurement (e.g. intensity)}

\item{cond}{character name of column with condition (e.g. Control, WT)}

\item{repl}{character name of column with replicate (e.g. unique experiment
identifiers)}

\item{pal}{name of colour palette to use (default is "tol_bright")}

\item{xlab}{string for x label (default is empty)}

\item{ylab}{string for y label (default is "Measurement")}

\item{datadist}{string for data distribution to use, select ("sina" default,
"jitter", or "violin")}

\item{size}{numeric vector of size range data and summary points (default is
c(0.8, 1.5))}

\item{alpha}{numeric vector of alpha range data and summary points (default
is c(0.5, 0.7))}

\item{bars}{string for type of error bars to add, select ("none" default,
"mean_sd", "mean_sem", or "mean_ci")}

\item{linking}{logical for whether to link summary points between conditions
(default is FALSE)}

\item{rep_summary}{string for summary statistic to use for replicates, select
("rep_mean" default, or "rep_median")}

\item{shapes}{logical for whether to use different shapes for replicates}

\item{fsize}{numeric font size for text (default is 9)}

\item{gg}{ggplot object to add to (default is NULL)}
}
\value{
ggplot object
}
\description{
Make a SuperPlot
}
\examples{
superplot(lord_jcb,
  "Speed", "Treatment", "Replicate",
  ylab = "Speed (um/min)"
)

}
