% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify_sp_columns}
\alias{verify_sp_columns}
\title{Verify the data frame used for SuperPlot}
\usage{
verify_sp_columns(df, meas, cond, repl)
}
\arguments{
\item{df}{data frame with at least three columns: meas, cond, repl}

\item{meas}{character name of column with measurement (e.g. intensity)}

\item{cond}{character name of column with condition (e.g. Control, WT)}

\item{repl}{character name of column with replicate (e.g. unique experiment
identifiers)}
}
\value{
logical to allow plot to go ahead
}
\description{
Verify the data frame used for SuperPlot
}
\keyword{internal}
